requirements_local.txt
-r requirements.txt
black
django-extensions
moto
pre-commit
pre-commit-hooks
pytest-cov
pytest-django
vcrpy
